pyproject.toml
[tool.poetry]
name = "jinja-docker-compose"
version = "1.1.2"
description = "Enables jinja like preposessing of docker-compose files."
homepage = "https://github.com/MorganLindqvist/jinja-compose/blob/master/README.md"
repository = "https://github.com/MorganLindqvist/jinja-compose"
authors = ["Morgan <Morgan@quzed.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "jinja_docker_compose", from = "src" }]

[tool.poetry.scripts]
jinja-docker-compose = 'jinja_docker_compose.main:cli'

[tool.poetry.dependencies]
python = "^3.7"
docker-compose = "^1.29.2"
Jinja2 = "^3.1.2"
PyYAML = "^5.4.1"
bcrypt = "^3.2.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
autopep8 = "^1.7.0"
pytest = "^7.1.3"
commitizen = "^2.35.0"
pre-commit = "^2.20.0"
tox = "^3.27.0"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.1.2"
tag_format = "v$version"
version_files = [
    "pyproject.toml:^version",
    "src/jinja_docker_compose/__version__.py",
]
annotated_tag = "true"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
